Review widget created
User can hover over a rating and the star and previous stars fill in yellow.
On click the star and previous stars change to green and the rating is saved.
On mouseenter the green disappears and the star and previous stars fill in yellow.
On mouseleave the yellow disappears and the saved green rating reappears.
A new rating can be made by clicking on a different star and the rating will update.