Added functionality to change the maximum number of stars that the review is out of.
This attribute updates on the click of the button and changes to the value of the input.
Set up a while loop so while index is less than the max-value attribute append a star in the #rating-container. Then increase the index by 1.

Finally the original methods for click, mouseenter and mouseleave have been modified to use delegate. This is so that the stars appended by updating the max value have the same functionality.