Updated text to be dynamic.
Why:
There were parts of the text which referenced out of 5 stars. If you have changed the max number of stars this number should reflect that.

How:
The number in the text was wrapped in span tags. Then during the click function the span was selected and the html altered to be the input value with id max-value.